chore: fixes dialog box component
