+++ src/js/components/Modals/AddFavouriteShabadModal.tsx
@@ -4,7 +4,7 @@ import Select from 'react-select';
 import { useCreateFavouriteShabad } from '@/components/FavouriteShabadButton/hooks';
 import { setModalOpen } from '@/features/actions';
 import { getShabadId } from '@/util';
-import Times from '@/components/Icons/Times';
+import Dialog from './Dialog';
 
 interface Props {
   isModalOpen: boolean;
@@ -34,48 +34,40 @@ const AddFavouriteShabadModal = (props: Props) => {
   }
 
   return (
-    <dialog open={props.isModalOpen} className='background-modal'>
-      <div className='add-fav-shabad'>
-        <div className='header'>
-          <span>Add to Favourites</span>
-          <button className='settings-times' aria-label="close" onClick={() => dispatch(setIsModalOpen(false))}><Times /></button>
-        </div>
-        <div className='content'>
-          <form method="dialog" onSubmit={handleShabadSave}>
-            <label className="title">Select a line to save as the title:
-              <Select
-                options={options}
-                defaultValue={options[0]}
-                className="dropdown"
-                classNamePrefix="react-select"
-                noOptionsMessage={() => null}
-                onChange={(val) => {
-                  if (val) {
-                    setPankti(val.value)
-                  }
-                }} />
+    <Dialog isModalOpen={props.isModalOpen} title="Add to Favourites">
+      <form method="dialog" onSubmit={handleShabadSave}>
+        <label className="title">Select a line to save as the title:
+          <Select
+            options={options}
+            defaultValue={options[0]}
+            className="dropdown"
+            classNamePrefix="react-select"
+            noOptionsMessage={() => null}
+            onChange={(val) => {
+              if (val) {
+                setPankti(val.value)
+              }
+            }} />
 
-            </label>
+        </label>
 
-            <label className='title'>Notes:
-              <textarea
-                name="comment"
-                rows={4}
-                cols={30}
-                style={{ resize: "none" }}
-                value={comment}
-                onChange={(e) => setComment(e.target.value)}
-                className="textarea"
-                autoFocus
-              />
-            </label>
-            <div className="save-btn">
-              <button type="submit" className='btn btn-primary'>Save</button>
-            </div>
-          </form>
+        <label className='title'>Notes:
+          <textarea
+            name="comment"
+            rows={4}
+            cols={30}
+            style={{ resize: "none" }}
+            value={comment}
+            onChange={(e) => setComment(e.target.value)}
+            className="textarea"
+            autoFocus
+          />
+        </label>
+        <div className="save-btn">
+          <button type="submit" className='btn btn-primary'>Save</button>
         </div>
-      </div>
-    </dialog>
+      </form>
+    </Dialog>
   )
 }
