+++ src/js/components/Modals/index.ts
@@ -1,1 +1,2 @@
 export { default as AddFavouriteShabadModal } from './AddFavouriteShabadModal';
+export { default as AskSinghBotQuestionModal } from './AskSinghBotQuestionModal';
