+++ src/js/util/misc/reformat-search-types.ts
@@ -2,8 +2,9 @@ import { SEARCH_TYPES } from '@/constants';
 
 // This is used for searchType options in HomePage and SearchPage
 export const reformatSearchTypes = (searchType: string[]) => {
-  const searchTypesObjArr = searchType.map((searchType, index) => ({ type: searchType, value: index }))
-
+  console.log(searchType,SEARCH_TYPES.ASK_A_QUESTION )
+  const searchTypesObjArr = searchType.map((searchTerm, index) => ({ type: searchTerm, value: index }))
+  console.log(searchTypesObjArr, 'search types obj arr')
   const mainLetterSearchIdx = SEARCH_TYPES.MAIN_LETTERS;
   const romanizedFirstLettersSearchIdx = SEARCH_TYPES.ROMANIZED_FIRST_LETTERS_ANYWHERE;
   const mainLetterSearchObj = searchTypesObjArr[mainLetterSearchIdx];
